# backend/src/API.yml
openapi: 3.0.3
info:
  description: This is the interactive documentation for the Innovatiehuis Intranet RESTful API
  version: "0.0.0"
  title: Innovatieplatform REST API
servers:
  - url: http://127.0.0.1:5000/api



  # Paths supported by the server application
paths:
...
  /projects/{project_id}:
    get:
      operationId: src.endpoints.projects.read_one
      tags:
        - Projects
      summary: Read one project
      parameters:
        - $ref: '#/components/parameters/project_id'
      responses:
        200:
          description: Successfully read project operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Project'
    put:
      operationId: src.endpoints.projects.update
      tags:
        - Projects
      summary: Update a value of project
      parameters:
        - $ref: '#/components/parameters/project_id'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                project_name:
                  type: string
                description:
                  type: string

      responses:
        200:
          description: Project updated
    delete:
      operationId: src.endpoints.projects.delete
      tags:
        - Projects
      summary: Deletes the specified project
      parameters:
        - $ref: '#/components/parameters/project_id'
      responses:
        200:
          description: Project deleted
...
  /announcements/{announcement_id}:
    get:
      operationId: src.endpoints.announcements.read_one
      tags:
        - Announcements
      summary: Reads a specific announcement
      parameters:
        - $ref: '#/components/parameters/announcement_id'
      responses:
        200:
          description: Successfully retrieved single announcement
    put:
      operationId: src.endpoints.announcements.update
      tags:
        - Announcements
      summary: Update title and/or content of a specific announcement
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Announcement'
      parameters:
        - $ref: '#/components/parameters/announcement_id'
      responses:
        200:
          description: Successfully posted announcement to project
    delete:
      operationId: src.endpoints.announcements.delete
      tags:
        - Announcements
      summary: Deletes specific announcement
      parameters:
        - $ref: '#/components/parameters/announcement_id'
      responses:
        200:
          description: Successfully removed announcement
...
  /users/{user_id}:
    get:
      operationId: src.endpoints.users.read_one
      tags:
        - Users
      summary: Get specified user
      parameters:
        - $ref: '#/components/parameters/user_id'
      responses:
        200:
          description: Successfully read user
    put:
      operationId: src.endpoints.users.update
      tags:
        - Users
      summary: Edit user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/User' # BUT NOT PASSWORD_HASH
      parameters:
        - $ref: '#/components/parameters/user_id'
      responses:
        200:
          description: Successfully edited user
        403:
          description: Request denied due to lack of permissions
    delete:
      operationId: src.endpoints.users.delete
      tags:
        - Users
      summary: Delete specified user
      parameters:
        - $ref: '#/components/parameters/user_id'
      responses:
        200:
          description: Successfully deleted user
  /users/{user_id}/screening/{screening_status}:
    patch:
      operationId: src.endpoints.users.update_screening
      tags:
        - Users
      summary: Change screening status of specified user
      parameters:
        - $ref: '#/components/parameters/user_id'
        - $ref: '#/components/parameters/screening_status'
      responses:
        200:
          description: Successfully updated password
        403:
          description: Request denied due to lack of permissions
  /users/{user_id}/role/{role_id}:
    patch:
      operationId: src.endpoints.users.update_role_user
      tags:
        - Users
        - Roles
      summary: Change screening status of specified user
      parameters:
        - $ref: '#/components/parameters/user_id'
        - $ref: '#/components/parameters/role_id'
      responses:
        200:
          description: Successfully updated password
        403:
          description: Request denied due to lack of permissions
  /users/{user_id}/password:
    patch:
      operationId: src.endpoints.users.update_password
      tags:
        - Users
      summary: Change password of specified user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: string
      parameters:
        - $ref: '#/components/parameters/user_id'
      responses:
        200:
          description: Successfully updated password
        403:
          description: Request denied due to lack of permissions
        404:
          description: Specified user not found
  /users/{user_id}/projects:
    get:
      operationId: src.endpoints.users.read_projects
      tags:
        - Users
      summary: Get projects of specified user
      parameters:
        - $ref: '#/components/parameters/user_id'
      responses:
        200:
          description: Successfully read user projects
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Project'
        404:
          description: Specified user not found
    post:
      operationId: src.endpoints.users.add_project
      tags:
        - Projects
      summary: Assign a new project to a user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Project'
      parameters:
        - $ref: '#/components/parameters/user_id'
      responses:
        200:
          description: Successfully assigned project to the user
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Project'
        404:
          description: Specified user or project not found
...
  /roles/{role_id}:
    put:
      operationId: src.endpoints.roles.update_role
      tags:
        - Roles
      summary: Update specified role
      parameters:
        - $ref: '#/components/parameters/role_id'
      responses:
        200:
          description: Successfully update role
    delete:
      operationId: src.endpoints.roles.delete_role
      tags:
        - Roles
      summary: Deletes specified role
      parameters:
        - $ref: '#/components/parameters/role_id'
      responses:
        200:
          description: Successfully deleted role
# ========================================AUTH/LOGOUT==================================================
...
components:
  parameters:
    project_id:
      in: path
      name: project_id
      required: true
      schema:
        type: string
      description: The ID number of a project
...
    user_id:
      in: path
      name: user_id
      required: true
      schema:
        type: string
      description: The ID number of a user
    announcement_id:
      in: path
      name: announcement_id
      required: true
      schema:
        type: string
      description: The ID number of an announcement
...
    role_id:
      in: path
      name: role_id
      required: true
      schema:
        type: string
      description: The ID number of a role
    screening_status:
      in: path
      name: screening_status
      required: true
      schema:
        type: string
      description: Number representing screening status; 0 = unscreened, 1 = in progress, 2 = screened
...
  schemas:
    Project:
      type: object
      properties:
        projectid:
          type: integer
        project_name:
          type: string
        is_archived:
          type: boolean
        create:
          type: string
        last_updated:
          type: string
    User:
      type: object
      properties:
        first_name:
          type: string
        last_name:
          type: string
        email:
          type: string
        phone_number:
          type: string
        roleid:
          type: integer
        screening_status:
          type: integer
        password_hash:
          type: string
    Announcement:
      type: object
      properties:
        content:
          type: string
        title:
          type: string
...